requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
streamlit>=1.35.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
requests>=2.31.0
plotly>=5.18.0
pytest>=9.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
